% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_flexible.R
\name{discrim_flexible}
\alias{discrim_flexible}
\alias{update.discrim_flexible}
\title{General Interface for Regularized Discriminant Models}
\usage{
discrim_flexible(mode = "classification", num_terms = NULL,
  prod_degree = NULL, prune_method = NULL)

\method{update}{discrim_flexible}(object, num_terms = NULL,
  prod_degree = NULL, prune_method = NULL, fresh = FALSE, ...)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "classification".}

\item{num_terms}{The number of features that will be retained in the
final model, including the intercept.}

\item{prod_degree}{The highest possible interaction degree.}

\item{prune_method}{The pruning method.}

\item{object}{A flexible discriminant model specification.}

\item{fresh}{A logical for whether the arguments should be
modified in-place of or replaced wholesale.}

\item{...}{Not used for \code{update()}.}
}
\description{
\code{discrim_flexible()} is a way to generate a \emph{specification} of a flexible
discriminant model using MARS features. The main arguments for the model are:
\itemize{
\item \code{num_terms}: The number of features that will be retained in the
final model.
\item \code{prod_degree}: The highest possible degree of interaction between
features. A value of 1 indicates and additive model while a value of 2
allows, but does not guarantee, two-way interactions between features.
\item \code{prune_method}: The type of pruning. Possible values are listed
in \code{?earth}.
}
}
\details{
These arguments are converted to their specific names at the
time that the model is fit. Other options and argument can be
set using \code{set_engine()}. If left to their defaults
here (\code{NULL}), the values are taken from the underlying model
functions. If parameters need to be modified, \code{update()} can be used
in lieu of recreating the object from scratch.

The model can be created using the \code{fit()} function using the
following \emph{engines}:
\itemize{
\item \pkg{R}:  \code{"earth"}  (the default)
}
}
\section{Engine Details}{


Engines may have pre-set default arguments when executing the
model fit call.  For this type of
model, the template of the fit calls are:

\pkg{rda}

\preformatted{
mda::fda(formula = missing_arg(), data = missing_arg(),
         nprune = num_terms, degree = prod_degree, pmethod = prune_method,
         method = earth::earth)
}
}

\examples{
discrim_flexible()
model <- discrim_flexible(num_terms = 10)
model
update(model, num_terms = 6)
update(model, num_terms = 6, fresh = TRUE)
}
