% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{klar_bayes_wrapper}
\alias{klar_bayes_wrapper}
\alias{pred_wrapper}
\title{Internal wrapper functions}
\usage{
klar_bayes_wrapper(x, y, ...)

pred_wrapper(object, new_data, ...)
}
\description{
Internal wrapper functions
}
\keyword{internal}
