% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrim_regularized.R
\name{rda_wrapper}
\alias{rda_wrapper}
\title{Helper functions for using parsnip with the klaR package}
\usage{
rda_wrapper(x, y, ...)
}
\description{
Helper functions for using parsnip with the klaR package
}
\keyword{internal}
