% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/naive_Bayes.R
\name{mass_lda_wrapper}
\alias{mass_lda_wrapper}
\alias{pred_wrapper}
\alias{klar_bayes_wrapper}
\title{Helper functions for using parsnip with the MASS package}
\usage{
mass_lda_wrapper(x, y, ...)

pred_wrapper(object, new_data, ...)

klar_bayes_wrapper(x, y, ...)
}
\description{
Helper functions for using parsnip with the MASS package
}
\keyword{internal}
